requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==4.5.1
SQLAlchemy==2.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.11.1
pytest==7.4.2
httpx==0.24.1
